Refactor how we obtain calendars and test Um Al Qura
