+++ src/NodaTime.Test/Calendars/PersianCalendarSystemTest.cs
@@ -21,11 +21,10 @@ namespace NodaTime.Test.Calendars
         public void BclThroughHistory()
         {
             Calendar bcl = BclCalendars.Persian;
-            // The "right" BCL equivalent to use depends on the version of .NET... pick it appropriately here.
-            CalendarSystem noda = PatternTestBase<LocalDate>.CalendarSystemForCalendar(bcl);
+            CalendarSystem noda = BclCalendars.CalendarSystemForCalendar(bcl);
             // Note: Noda Time stops in 9377, whereas the BCL goes into the start of 9378. This is because
             // Noda Time ensures that the whole year is valid.
-            BclEquivalenceHelper.AssertEquivalent(bcl, noda, noda.MinYear, noda.MaxYear);
+            BclEquivalenceHelper.AssertEquivalent(bcl, noda);
         }
 
         /// <summary>
